- moved groups page to a component to fix header issue on homepage - logo and title only showed on component not on homepage
